Added PSR4 autoloader and separated the code with Tree and Node classes. index.php is used for test purposes.
